Rename file readfirstlance.hpp as amd_wave_read_first_lane.hpp
