+++ include/ck/utility/amd_wave_read_first_lane.hpp
